refactor: move subgraph-specific-libs to lib/subgraph
