+++ apps/ensindexer/src/lib/subgraph/is-label-subgraph-indexable.ts
