+++ apps/ensindexer/src/plugins/subgraph/shared-handlers/Registrar.ts
@@ -17,12 +17,12 @@ import {
 } from "@ensnode/ensnode-sdk";
 
 import config from "@/config";
-import { sharedEventValues, upsertAccount, upsertRegistration } from "@/lib/db-helpers";
 import { labelByLabelHash } from "@/lib/graphnode-helpers";
-import { makeRegistrationId } from "@/lib/ids";
-import { isLabelSubgraphIndexable } from "@/lib/is-label-subgraph-indexable";
 import { pluginSupportsPremintedNames } from "@/lib/plugin-helpers";
 import type { EventWithArgs } from "@/lib/ponder-helpers";
+import { sharedEventValues, upsertAccount, upsertRegistration } from "@/lib/subgraph/db-helpers";
+import { makeRegistrationId } from "@/lib/subgraph/ids";
+import { isLabelSubgraphIndexable } from "@/lib/subgraph/is-label-subgraph-indexable";
 import type { RegistrarManagedName } from "@/lib/types";
 import { handleNewOwner } from "@/plugins/subgraph/shared-handlers/Registry";
