+++ apps/ensindexer/test/subgraph/ids.test.ts
@@ -2,7 +2,7 @@ import { beforeEach, describe, expect, it } from "vitest";
 import { resetMockConfig, setupConfigMock, updateMockConfig } from "./utils/mockConfig";
 setupConfigMock(); // setup config mock before importing dependent modules
 
-import { makeEventId, makeRegistrationId, makeResolverId } from "@/lib/ids";
+import { makeEventId, makeRegistrationId, makeResolverId } from "@/lib/subgraph/ids";
 import { labelhash, namehash, zeroAddress } from "viem";
 
 const CHAIN_ID = 1337;
